+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="co.devskills.mobile_boilerplate_android_java">
+    package="io.alvalabs.mobile_boilerplate_android_java">
 
     <application
         android:allowBackup="true"
