+++ app/src/main/java/io/alvalabs/mobile_boilerplate_android_java/ui/dashboard/DashboardViewModel.java
@@ -1,16 +1,16 @@
-package co.devskills.mobile_boilerplate_android_java.ui.notifications;
+package io.alvalabs.mobile_boilerplate_android_java.ui.dashboard;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class DashboardViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public DashboardViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is dashboard fragment");
     }
 
     public LiveData<String> getText() {
