+++ app/src/main/java/io/alvalabs/mobile_boilerplate_android_java/ui/home/HomeViewModel.java
@@ -1,16 +1,16 @@
-package co.devskills.mobile_boilerplate_android_java.ui.dashboard;
+package io.alvalabs.mobile_boilerplate_android_java.ui.home;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class HomeViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public HomeViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is home fragment");
     }
 
     public LiveData<String> getText() {
