+++ app/src/main/java/io/alvalabs/mobile_boilerplate_android_java/ui/notifications/NotificationsViewModel.java
@@ -1,16 +1,16 @@
-package co.devskills.mobile_boilerplate_android_java.ui.home;
+package io.alvalabs.mobile_boilerplate_android_java.ui.notifications;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class NotificationsViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public NotificationsViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is notifications fragment");
     }
 
     public LiveData<String> getText() {
