+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/navigation_home"
-        android:name="co.devskills.mobile_boilerplate_android_java.ui.home.HomeFragment"
+        android:name="io.alvalabs.mobile_boilerplate_android_java.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="co.devskills.mobile_boilerplate_android_java.ui.dashboard.DashboardFragment"
+        android:name="io.alvalabs.mobile_boilerplate_android_java.ui.dashboard.DashboardFragment"
         android:label="@string/title_dashboard"
         tools:layout="@layout/fragment_dashboard" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="co.devskills.mobile_boilerplate_android_java.ui.notifications.NotificationsFragment"
+        android:name="io.alvalabs.mobile_boilerplate_android_java.ui.notifications.NotificationsFragment"
         android:label="@string/title_notifications"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
